Add a pricing collector
As a preparation to make #15 possible I implemented a pricing collector.
It exposes the following metrics including gross and net prices:

* hcloud_pricing_floating_ip
  The cost of one floating IP per month

* hcloud_pricing_image
  The cost of an image per GB/month

* hcloud_pricing_server_backup
  Will increase base server costs by specific percentage if server backups are enabled

* hcloud_pricing_traffic
  The cost of additional traffic per TB

The Hetzner cloud API also exposes prices for all server types in all
regions. It would be possible to collect those as well but it would
produce a lot of data with limited use. (Every server type is available
with two different storage technologies and in three different
regions.) If someone needs all thoses prices it should be easy to add in
the future.

Signed-off-by: Max Rosin <git@hackrid.de>