Fix out of range slice access in server metrics
The code assumed there to always be a single metrics datapoint returned
from the API. In some cases this is not true though, and an index out of
range error would occur.

This code explicitly checks for at least one datapoint to be present,
and use the last returned datapoint. It should handle omitted datapoints
more gracefully.